requirements.txt
Flask==3.1.0
flask-cors==5.0.0
Werkzeug==3.1.3
python-dotenv==1.0.1
requests==2.32.3
openai==1.61.0
python-docx==1.1.2
groq==0.17.0
pydub==0.25.1
boto3
botocore
